Fix server signature
